Refactored OIDE to be run as a per-user process.
