+++ oide/lib/handlers/pam_auth.py
@@ -1,5 +1,5 @@
-import os
-import Pyro4
+import logging
+from simplepam import authenticate
 from oide.lib.handlers.base import BaseHandler
 
 import oide.settings as global_settings
@@ -14,14 +14,8 @@ class PAMLoginHandler(BaseHandler):
     def post(self):
         un = self.get_argument('username')
         pw = str(self.get_argument('password'))
-        auth_pam = Pyro4.Proxy('PYRONAME:%s@%s:%d'%(
-            global_settings.PYRO_AUTHMODULE_URI,
-            global_settings.PYRO_NAMESERVER_HOST,
-            global_settings.PYRO_NAMESERVER_PORT
-            )
-        )
 
-        if auth_pam.authenticate(un, pw):
+        if authenticate(un, str(pw), service='login'):
             self.set_secure_cookie('user', un)
             self.redirect("/")
         else:
