--- oide/lib/services/__init__.py
@@ -1,5 +1,0 @@
-import os
-import glob
-
-modules = glob.glob(os.path.dirname(__file__)+"/*.py")
-__all__ = [ os.path.basename(f)[:-3] for f in modules]
